Handle asynchronous result properly
On a first glance it wasn't clear that the call was async and that we
should resolve the Futures before retuning to users.

I decided to wait for 30 seconds as that is the timeout for publishing
and I guess used to be the default timeout for all other calls.

I also handled interrupt exception properly.