Use token based instead of offset based approach for subscriptions listing
In case if there are a lot of subscriptions in the database, using offset
based approach for pagination is not really something that works at scale.
Therefore - there is an attempt to replace offset approach with token
approach, that is much more widely adopted and may actually solve the
problem with db overusage.

Token in this case is a string, representing subscription_id and direction
of iteration.
In case if this approach will be accepted by the nakadi team - tests
will be written in order to actually test the behavior (there is a bug
right now with backward iteration to the first page in case if
subscription was removed during iteration, but I am on vacation and we
are treating it as external contribution)