Default statistics backward compatible change
In the PR where we changed the repartitioning endpoint we also
accidentally introduced a change in an undocumented behaviour that
some clients relied on. It used to be possible to update an event type
without providing default statistics.

We believe this behaviour should not be modified because this field is
optional, so some clients might not even support it (Nakadi UI, for
example).

We are also fixing a problem where repartition could be triggered
concurrently with an event type update resulting in some race
condition on database and timeline changes in database.