[WIP] Change cleanup policy from delete to compact_and_delete
This commit is just a work in progress. It allows to change the
cleanup policty from delete to compact_and_delete.

On the way to implement this, I had to read some code that I could not
avoid changing. But I'm not sure it's looking better either.

I removed one unused feature toggle.

We need to figure out what to do with retention time that is null or
different from the actual config in Kafka.

WIP: change cleanup policy

[WIP] Change cleanup policy from delete to compact_and_delete

This commit is just a work in progress. It allows to change the
cleanup policty from delete to compact_and_delete.

On the way to implement this, I had to read some code that I could not
avoid changing. But I'm not sure it's looking better either.

I removed one unused feature toggle.

We need to figure out what to do with retention time that is null or
different from the actual config in Kafka.