Remove logging and stop loosing log messages.
There is inoptimal thread usage for kpi and supplementary mesage publishing.
One or several threads may actually block whole the processing flow.
In order to avoid it intermediate batcher thread was introduced, with the
only purpose of batching requests.
The fact that it may not catch up with creating threads is clear,
and will be addressed in next commit.