Building and running on Java 11 (#1248)
* Building and running on Java 11

This PR makes the least changes necessary to build and run Nakadi on
Java 11.

- uses approved base image
https://cloud.docs.zalando.net/howtos/compliant-images/#use-of-allowed-docker-base-images
- upgraded gradle to 6.7.1 which introduced `toolchain`
https://docs.gradle.org/current/userguide/toolchains.html which helps
us make sure that the same java version is used everywhere
- included an extra dependency (unfortunatelly) to fix springboot
dependency which is no longer part of default java module
https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-with-Java-9-and-above

This way the app builds and passes acceptance tests. I'm not sure this
is production ready yet as there could be changes in GC, for
example. Some work is still required.

* Using overlay for java 11

This may speed up build time as gradle would identify the available
java 11 in the machine and not download it before building.