use optimistic locking to upload subscription topology (#1277)
* use optimistic locking to upload subscription topology

the commit changes approach on how to do distributed locking in favour of optimistic locking. the current approach with apache curator lock recipe brings issues like not all reassignments are done in time which introduces errors while openning subscriptions or doing client rebalancing. the new approach is leveraging zookeeper version to upload changes if version is changed it will read changed structure from zookeeper and if the change still requires it will retry with the new version.