better track inability to join streaming on consumption
In order to synchronize behavior of several instances, there is an optimistic lock is in use.
Unfortunately one place was not handled properly - registration of the streaming session.
The problem is that streaming code is allowing to have more streaming sessions, than there are partitions, or have more than 1 streaming connections for manually assigned partition.
This causes some problems - after unsuccessful rebalance (obviously rebalance is saying that it can't stream 2 partitions to 3 connections) some of streaming connections are dying.

In order to avoid that, there should be 2 changes made:
1. Session should be registered in optimistic manner (check, register, check. If not - session must be removed, https://github.com/zalando/nakadi/blob/master/api-consumption/src/main/java/org/zalando/nakadi/service/subscription/state/StartingState.java#L46-L87)
2. It should be expected, that rebalance may predictably fail for a very short amount of time, and if it is - it should retry (usage of https://github.com/zalando/nakadi/blob/master/api-consumption/src/main/java/org/zalando/nakadi/service/subscription/SubscriptionRebalancer.java).