part 1: publishing to nakadi access log in avro (#1359)
* internal event type publishing

Envelope

Nakadi deserializes event’s metadata to identify Kafka partitions to publish, event compaction key and extends metadata with additional fields like event type, publisher uid, event receive time and JSON schema version. Avro is serialized to binary format. In order for Nakadi to deserialize Avro bytes it has to know the exact position of metadata. Hence, the envelope format has to be positional.


Field name	Metadata Version	Metadata length	Metadata	Payload length	Payload
Size (bytes)	1	                        4	                         x	                4	                         y
Description	Schema version to parse metadata	Metadata size in bytes	Message metadata	Payload size in bytes	User payload

The format works with the single messages as well as with an array of messages in a single request. The overall structure of the envelope does not depend on the format of the payload. This means, new formats can be used for payload in the future.