The first part in the series of changes to support binary formats in Nakadi (#1362)
The first part in the series of changes to support binary formats in Nakadi, in that particular reincarnation is Apache Avro.

Apache Avro™ is a data serialization system. Avro provides a compact, fast, binary data format and simple integration with dynamic languages. Avro supports code generation as an additional optimization, but it is only worth utilizing for statically typed languages. According to the research, it is faster in serialization and deserialization and around 10% more compact than JSON.

Nakadi publishes Nakadi KPI event type nakadi.access.log in Avro to Kafka using the following format for Kafka record:
Nakadi deserializes event’s metadata to identify Kafka partitions to publish, event compaction key and extends metadata with additional fields like event type, publisher uid, event receive time and JSON schema version. Avro is serialized to binary format. In order for Nakadi to deserialize Avro bytes it has to know the exact position of metadata. Hence, the envelope format has to be positional.
Field name	Metadata Version	Metadata length	Metadata	Payload length	Payload
Size (bytes)	1	4	x	4	y
Description	Schema version to parse metadata	Metadata size in bytes	Message metadata	Payload size in bytes	User payload
The format works with the single messages as well as with an array of messages in a single request. The overall structure of the envelope does not depend on the format of the payload. This means, new formats can be used for payload in the future.

Consumption of the Avro serialised events happens in JSON using Subscription API

At the moment Avro schemas for Nakadi metadata and event are stored in Nakadi source code. The schema registry will adopt it in the future.

The changes were reviewed in the following split PRs: #1358, #1359, #1360

Co-authored-by: a1exsh <a1exsh@users.noreply.github.com>