Introducing KPIEvent and supporting Avro for nakadi.subscription.log
A new class called `KPIEvent` is introduced. All KPI events are supposed
to be exted from this. Currently the only KPI event that is published
using this is the `SubscriptionLogEvent`. The `KPIEvent` supports
mapping the classes that extends it directly to `GenericRecord` or
`JSONObject`; as long as the fields are annotated with `@KPIField`
annotation. The `NakadiKpiPublisher` class is also supporting the
publishing of KPI Events now.